requirements.txt
cycler==0.10.0
matplotlib==1.5.1
numpy==1.11.0
pandas==0.18.1
Pillow==3.2.0
protobuf==3.0.0b2
pyparsing==2.1.4
python-dateutil==2.5.3
pytz==2016.4
scikit-learn==0.17.1
scipy==0.17.1
seaborn==0.7.0
six==1.10.0
tensorflow==0.8.0
wheel==0.29.0
